+++ tasks/CheckFlagAsCriticalSR.php
@@ -11,6 +11,7 @@
 use CNCLTD\LoggerCLI;
 
 require_once(__DIR__ . "/../htdocs/config.inc.php");
+global $cfg;
 require_once($cfg["path_dbe"] . "/DBEProblem.inc.php");
 require_once($cfg['path_bu'] . '/BUHeader.inc.php');
 global $db;