+++ tasks/CheckPendingCompletionSR.php
@@ -1,13 +1,4 @@
 <?php
-
-
-/**
- * Created by PhpStorm.
- * User: fizdalf
- * Date: 17/12/2018
- * Time: 11:26
- */
-
 use CNCLTD\LoggerCLI;
 use Twig\Environment;