+++ tasks/NonUKHolidaysEmails.php
@@ -1,24 +1,33 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: fizdalf
- * Date: 13/04/2018
- * Time: 12:28
- */
+use CNCLTD\LoggerCLI;
+require_once(__DIR__ . "/../htdocs/config.inc.php");
+global $cfg;
 
-/**
- * Check that the mail queue has no emails older than 15 minutes
- *
- * If it does then email graham and gary and Karim
- *
- * @authors Karim Ahmed - Sweet Code Limited
- */
-require_once("config.inc.php");
+$logName = 'NonUKHolidaysEmails';
+$logger = new LoggerCLI($logName);
+
+// increasing execution time to infinity...
+ini_set('max_execution_time', 0);
+
+if (!is_cli()) {
+    echo 'This script can only be ran from command line';
+    exit;
+}
+// Script example.php
+$shortopts = "d";
+$longopts = [
+    "testDate:"
+];
+$options = getopt($shortopts, $longopts);
+$debugMode = false;
+if (isset($options['d'])) {
+    $debugMode = true;
+}
 global $cfg;
 require_once($cfg["path_bu"] . "/BUMail.inc.php");
 require_once($cfg ["path_func"] . "/Common.inc.php");
 
-$dateToTest = @$_REQUEST['testDate'];
+$dateToTest = @$options['testDate'];
 
 $error = false;