+++ tasks/TeamPerformanceUpdate.php
@@ -1,6 +1,26 @@
 <?php
-require_once("config.inc.php");
+use CNCLTD\LoggerCLI;
+require_once(__DIR__ . "/../htdocs/config.inc.php");
 global $cfg;
+
+$logName = 'TeamPerformanceUpdate';
+$logger = new LoggerCLI($logName);
+
+// increasing execution time to infinity...
+ini_set('max_execution_time', 0);
+
+if (!is_cli()) {
+    echo 'This script can only be ran from command line';
+    exit;
+}
+// Script example.php
+$shortopts = "d";
+$longopts = [];
+$options = getopt($shortopts, $longopts);
+$debugMode = false;
+if (isset($options['d'])) {
+    $debugMode = true;
+}
 require_once($cfg["path_bu"] . "/BUTeamPerformance.inc.php");
 $thing = null;
 $buTeamPerformance = new BUTeamPerformance($thing);